Refactored framework code to be in separate files
